feat: fab에footer intersecting 기능 추가
