+++ components/AddLinkField/AddLinkField.tsx
@@ -1,13 +1,6 @@
 "use client";
 
-import {
-  ForwardedRef,
-  forwardRef,
-  useCallback,
-  useEffect,
-  useRef,
-  useState,
-} from "react";
+import { ForwardedRef, forwardRef } from "react";
 
 import { IFolder, ILink } from "@/types/linkbrary";
 
@@ -19,48 +12,28 @@ interface IAddLinkFieldProps {
   folders: IFolder[] | [];
   links: ILink[] | [];
   inView: boolean | null;
+  isTransition: boolean;
+  scrollRef: React.MutableRefObject<HTMLDivElement | null>;
 }
 
 const AddLinkField = forwardRef(function AddLinkField(
-  { userId, folders, links, inView }: IAddLinkFieldProps,
+  {
+    userId,
+    folders,
+    links,
+    inView,
+    isTransition,
+    scrollRef,
+  }: IAddLinkFieldProps,
   ref: ForwardedRef<HTMLDivElement>
 ) {
-  const FOOTER_HEIGHT = 160;
-  const scrollRef = useRef<HTMLDivElement>(null);
-  const [throttle, setThrottle] = useState(false);
-  const [transition, setTransition] = useState(false);
-
-  const handleScroll = useCallback(() => {
-    if (throttle) return;
-    if (!throttle) {
-      setThrottle(true);
-      setTimeout(() => {
-        if (scrollRef.current) {
-          const currentScrollPosition =
-            scrollRef.current.getBoundingClientRect().bottom + window.scrollY; // pageYOffset == scrollY
-          const IntersectedFooter = document.body.scrollHeight - FOOTER_HEIGHT;
-          setTransition(currentScrollPosition >= IntersectedFooter);
-          setThrottle(false);
-        }
-      }, 300);
-    }
-  }, [throttle]);
-
-  useEffect(() => {
-    handleScroll();
-    window.addEventListener("scroll", handleScroll);
-    return () => {
-      window.removeEventListener("scroll", handleScroll);
-    };
-  }, [handleScroll]);
-
   return (
     <div className={styles.observedWrapper}>
       <div
         ref={scrollRef}
         className={`${styles.addLinkContainer} ${
           styles[`${inView ? "view" : "not-view"}`]
-        } ${styles[`${transition ? "tr" : "not-tr"}`]}`}
+        } ${styles[`${isTransition ? "tr" : "not-tr"}`]}`}
       >
         <AddLinkBar userId={userId} folders={folders} links={links} />
       </div>
