+++ components/FolderContents/FolderContents.tsx
@@ -1,5 +1,7 @@
 "use client";
 
+import { useCallback, useEffect, useRef, useState } from "react";
+
 import CurrentFolderMenu from "@/components/CurrentFolderMenu/CurrentFolderMenu";
 import FolderChipField from "@/components/FolderChipField/FolderChipField";
 import LinkField from "@/components/LinkField/LinkField";
@@ -31,6 +33,35 @@ const FolderContents = ({
     : folders;
   const currentFolder = folderList.find((folder) => folder.id === currentTab);
 
+  const FOOTER_HEIGHT = 160;
+  const scrollRef = useRef<HTMLDivElement>(null);
+  const [throttle, setThrottle] = useState(false);
+  const [isTransition, setIsTransition] = useState(false);
+
+  const handleScroll = useCallback(() => {
+    if (throttle) return;
+    if (!throttle) {
+      setThrottle(true);
+      setTimeout(() => {
+        if (scrollRef.current) {
+          const currentScrollPosition =
+            scrollRef.current.getBoundingClientRect().bottom + window.scrollY; // pageYOffset == scrollY
+          const IntersectedFooter = document.body.scrollHeight - FOOTER_HEIGHT;
+          setIsTransition(currentScrollPosition >= IntersectedFooter);
+          setThrottle(false);
+        }
+      }, 300);
+    }
+  }, [throttle]);
+
+  useEffect(() => {
+    handleScroll();
+    window.addEventListener("scroll", handleScroll);
+    return () => {
+      window.removeEventListener("scroll", handleScroll);
+    };
+  }, [handleScroll]);
+
   return (
     <>
       <AddLinkField
@@ -38,6 +69,8 @@ const FolderContents = ({
         folders={folders}
         links={links}
         inView={inView}
+        isTransition={isTransition}
+        scrollRef={scrollRef}
         ref={(el: HTMLDivElement) => (observerTargetRefs.current[0] = el)}
       />
       <div className={styles.contents}>
@@ -48,6 +81,7 @@ const FolderContents = ({
           currentTab={currentTab}
           inView={inView}
           isLinks={links.length !== 0}
+          isTransition={isTransition}
         />
         {currentFolder && <CurrentFolderMenu currentFolder={currentFolder} />}
         <LinkField userId={userId} folders={folders} links={links} />
